Refactor: Put network utils in their own file
