+++ scaler/io/ymq/error.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <algorithm>
-#include <cstring>
 #include <exception>  // std::terminate
 #include <format>
 #include <functional>
@@ -9,26 +8,7 @@
 #include <string>
 
 #include "scaler/io/ymq/timestamp.h"
-
-template <std::size_t N>
-    requires(N > 0)
-consteval auto getFormatString() {
-    std::string str = "{}";
-    for (size_t i = 1; i < N; ++i)
-        str += ": {}";
-    std::array<char, (N - 1) * 4 + 2> arr;
-    std::ranges::copy(str, arr.begin());
-    return arr;
-}
-
-// NOTE: It seems like this two lines cannot be placed in the constructor for unknown reason.
-template <typename... Args>
-constexpr std::string argsToString(Args&&... args) {
-    static constexpr const auto str = getFormatString<sizeof...(Args)>();
-
-    std::string res = std::format(std::string_view {str}, std::forward<Args>(args)...);
-    return res;
-}
+#include "scaler/io/ymq/utils.h"
 
 struct Error: std::exception {
     enum struct ErrorCode {
